+++ Libplanet/Store/IStore.cs
@@ -42,6 +42,17 @@ namespace Libplanet.Store
         /// <seealso cref="GetCanonicalChainId()"/>
         void SetCanonicalChainId(Guid chainId);
 
+        /// <summary>
+        /// Returns the genesis block of the current canonical chain.
+        /// </summary>
+        /// <param name="hashAlgorithmGetter">The function to determine hash algorithm used for
+        /// proof-of-work mining.</param>
+        /// <typeparam name="T">An <see cref="IAction"/> type.  It should match
+        /// to <see cref="Block{T}"/>'s type parameter.</typeparam>
+        /// <returns>The genesis block of the current canonical chain.</returns>
+        Block<T> GetCanonicalGenesisBlock<T>(HashAlgorithmGetter hashAlgorithmGetter)
+            where T : IAction, new();
+
         long CountIndex(Guid chainId);
 
         /// <summary>